Now the SearchBar is global
